requirements/base.txt
amqp == 1.4.6
anyjson == 0.3.3
argh == 0.24.1
billiard == 3.3.0.20
bleach == 1.4
boto == 2.27.0
celery == 3.1.18
dj-database-url == 0.3.0
Django == 1.8.2
djangorestframework == 3.1.3
django-allauth == 0.21.0
django-bootstrap3 == 3.0.0
django-braces == 1.3.1
django-celery == 3.1.16
django-debug-toolbar == 1.3.2
django-extensions == 1.3.3
django-filter == 0.9.1
django-fsm == 1.5.1
django-model-utils == 2.0.2
django-object-actions == 0.4.0
django-redis == 4.2.0
django-storages == 1.1.8
django-suit == 0.2.13
django-vanilla-views == 1.0.2
django-waffle == 0.10.1
docopt == 0.4.0
envdir == 0.6.1
feedparser == 5.1.3
html5lib == 0.999
ipython == 2.0.0
Jinja2 == 2.7.2
kombu == 3.0.26
logutils == 0.3.3
mailchimp == 2.0.7
Markdown == 2.5.2
meld3 == 0.6.10
msgpack-python == 0.4.6
oauthlib == 0.7.2
pathtools == 0.1.2
psycopg2 == 2.5.2
Pygments == 1.6
python-openid == 2.2.5
pytz == 2015.4
PyYAML == 3.10
pyzmq == 14.1.1
redis == 2.10.3
requests == 2.7.0
requests-oauthlib == 0.5.0
rollbar == 0.9.4
setuptools == 3.4.3
sqlparse == 0.1.15
supervisor == 3.0
tornado == 3.2
Unipath == 1.0
watchdog == 0.7.1

# django-supervisor == 0.3.2
-e git+https://github.com/fanscribed/django-supervisor@7c6d0d0b48#egg=django-supervisor
